refactor: extract shared render path to core::render (issue 421)
Desktop and test-bench duplicated the offscreen render pipeline +
capture readback. ADR-0067 acknowledged it; this lifts the shared
shape into `aether-substrate-core::render`, gated by a `render`
feature so headless and hub still don't pull wgpu/png.

The new core::render exposes:
- Pipeline (offscreen draw pipeline + camera bind group + buffers)
- Targets (offscreen color + depth + lazy readback)
- record_main_pass (uploads + draws, accepts extra pipelines for
  desktop's wireframe overlay)
- prepare_capture_copy + finish_capture (the BGRA-aware readback
  path; test-bench passes RGBA, desktop passes whatever the surface
  negotiated)
- shader.wgsl (consolidated to one copy)
- The shared constants and vertex_buffer_layout()

Each chassis still owns its own `Gpu` struct as the chassis-shaped
wrapper. Desktop adds surface acquisition, swapchain blit, present,
and the optional wireframe overlay; test-bench is just the offscreen
path. Both delegate the byte-level work to core::render.

Verified wgpu absent from headless + hub + core's default-feature
trees via cargo tree; present in desktop + test-bench via the render
feature.

Closes #421
