--- crates/aether-substrate-test-bench/src/shader.wgsl
@@ -1,36 +1,0 @@
-// Vertex positions are world-space (x, y, z). The camera uniform
-// `view_proj` (column-major, uploaded verbatim from the
-// `aether.camera` sink) multiplies every vertex to produce clip
-// space. Before the first `Camera` mail arrives the uniform holds
-// identity, so components still emitting clip-space-ish world coords
-// render unchanged.
-
-struct Camera {
-    view_proj: mat4x4<f32>,
-}
-
-@group(0) @binding(0)
-var<uniform> camera: Camera;
-
-struct VertexInput {
-    @location(0) position: vec3<f32>,
-    @location(1) color: vec3<f32>,
-}
-
-struct VertexOutput {
-    @builtin(position) clip_pos: vec4<f32>,
-    @location(0) color: vec3<f32>,
-}
-
-@vertex
-fn vs_main(in: VertexInput) -> VertexOutput {
-    var out: VertexOutput;
-    out.clip_pos = camera.view_proj * vec4<f32>(in.position, 1.0);
-    out.color = in.color;
-    return out;
-}
-
-@fragment
-fn fs_main(in: VertexOutput) -> @location(0) vec4<f32> {
-    return vec4<f32>(in.color, 1.0);
-}
